e2e revails: Call without args fails
