+++ e2e-tests/tests/call.rs
@@ -1,0 +1,43 @@
+use ic_cdk_e2e_tests::cargo_build_canister;
+use pocket_ic::common::rest::RawEffectivePrincipal;
+use pocket_ic::{call_candid, PocketIcBuilder};
+
+#[test]
+fn call_echo() {
+    let pic = PocketIcBuilder::new()
+        .with_application_subnet()
+        .with_nonmainnet_features(true)
+        .build();
+    let wasm = cargo_build_canister("call");
+    let canister_id = pic.create_canister();
+    pic.add_cycles(canister_id, 100_000_000_000_000);
+    pic.install_canister(canister_id, wasm, vec![], None);
+    let _: () = call_candid(
+        &pic,
+        canister_id,
+        RawEffectivePrincipal::None,
+        "call_echo",
+        (),
+    )
+    .expect("Error calling call_echo");
+}
+
+#[test]
+fn call_foo() {
+    let pic = PocketIcBuilder::new()
+        .with_application_subnet()
+        .with_nonmainnet_features(true)
+        .build();
+    let wasm = cargo_build_canister("call");
+    let canister_id = pic.create_canister();
+    pic.add_cycles(canister_id, 100_000_000_000_000);
+    pic.install_canister(canister_id, wasm, vec![], None);
+    let _: () = call_candid(
+        &pic,
+        canister_id,
+        RawEffectivePrincipal::None,
+        "call_foo",
+        (),
+    )
+    .expect("Error calling call_foo");
+}
